refactor(core): extract runtime service configuration
