+++ framework/maintainability/abstraction-integrity-report.json
@@ -6,14 +6,14 @@
   "measurement": {
     "aggregates": {
       "kungfu-cli": 20273,
-      "kungfu-flat-root": 35569,
-      "kungfu-runtime-family": 6160
+      "kungfu-flat-root": 35651,
+      "kungfu-runtime-family": 6242
     },
     "counts": {
-      "productionFiles": 233,
-      "productionPhysicalLines": 83458,
+      "productionFiles": 234,
+      "productionPhysicalLines": 83540,
       "testFiles": 97,
-      "testPhysicalLines": 45648
+      "testPhysicalLines": 45663
     },
     "hiddenProductionFiles": [],
     "modulesOver1000": [
@@ -87,7 +87,7 @@
       },
       {
         "path": "framework/core/src/python/kungfu/runtime_service.py",
-        "physicalLines": 2303
+        "physicalLines": 2000
       },
       {
         "path": "framework/core/src/python/kungfu/runtime_upgrade.py",
@@ -114,10 +114,6 @@
       {
         "path": "framework/core/src/python/kungfu/profile_sdk.py",
         "physicalLines": 3135
-      },
-      {
-        "path": "framework/core/src/python/kungfu/runtime_service.py",
-        "physicalLines": 2303
       },
       {
         "path": "framework/core/src/python/kungfu/workspace_federation.py",
@@ -228,10 +224,10 @@
         "responsibilities": 114
       },
       {
-        "contentRoot": "sha256:596b1331379e6c071e84faee38371a4d0135154dcdf7358bea977cfac918aed5",
+        "contentRoot": "sha256:1dd95c5a35bccbdd036f91169445cc29c2d1ed4b1bcb3baf9292d3af4c54700b",
         "path": "framework/core/src/python/kungfu/runtime_service.py",
-        "physicalLines": 2303,
-        "responsibilities": 104
+        "physicalLines": 2000,
+        "responsibilities": 90
       },
       {
         "contentRoot": "sha256:6209d0c4c7566a0f4dce029483972f3e58136aff1c4ec15ad2ec30791c2ddad9",
@@ -341,17 +337,17 @@
         "test": "platform.system() != 'Windows'"
       },
       {
-        "line": 1523,
+        "line": 1371,
         "path": "framework/core/src/python/kungfu/runtime_service.py",
         "test": "platform.system() == 'Windows'"
       },
       {
-        "line": 1543,
+        "line": 1391,
         "path": "framework/core/src/python/kungfu/runtime_service.py",
         "test": "platform.system() != 'Windows' or getattr(error, 'winerror', None) != 5"
       },
       {
-        "line": 1631,
+        "line": 1479,
         "path": "framework/core/src/python/kungfu/runtime_service.py",
         "test": "platform.system() != 'Windows'"
       },
@@ -362,7 +358,7 @@
       }
     ],
     "schema": "kungfu.python-structure-measurement/v1",
-    "sourceRoot": "sha256:c9fea7463dd1344acf5d42c3819e28f71b22087437afdb8cc919d2e6db89741c",
+    "sourceRoot": "sha256:0cbcbcd49a89e39d9151bd577d48a7024047a52ee670c30d9a651f325af37526",
     "sourceRoots": {
       "production": [
         "framework/core/src/python"
@@ -425,6 +421,12 @@
       ]
     },
     {
+      "owner": "core/python-runtime-service-config",
+      "paths": [
+        "framework/core/src/python/kungfu/runtime_service_config.py"
+      ]
+    },
+    {
       "owner": "shifu/python-structure-entry",
       "paths": [
         "scripts/check-code-complexity.mjs",
@@ -454,9 +456,9 @@
     "kungfu.runtime_broker.NativeReadinessAuthority",
     "kungfu.runtime_service.ProcessRuntimeHost"
   ],
-  "reportRoot": "sha256:9bad76dea6c4d806a85a0fca9a63ccc018d403aa1cb26f2fe1aaa746348d3dd0",
+  "reportRoot": "sha256:47fee75748e976feac0d65fc661ac16ecf9c136a6b75256c2545e1723a869858",
   "schema": "kungfu.abstraction-integrity-report/v1",
-  "sourceRoot": "sha256:c9fea7463dd1344acf5d42c3819e28f71b22087437afdb8cc919d2e6db89741c",
+  "sourceRoot": "sha256:0cbcbcd49a89e39d9151bd577d48a7024047a52ee670c30d9a651f325af37526",
   "summary": {
     "blockingIssues": 0
   },
